refactor: move isSupportedForRAG and getSupportedFileTypes functions to utils for better organization
